Backend

+++ src/main/java/com/codewitharjun/fullstackbackend/exception/TarifNotFoundException.java
@@ -2,7 +2,7 @@ package com.codewitharjun.fullstackbackend.exception;
 
 public class TarifNotFoundException extends RuntimeException{
 
-    public TarifNotFoundException(Long nomenclature){
+    public TarifNotFoundException(Integer nomenclature){
         super("Could not found the Tarif with nomenclature "+ nomenclature);
     }
 }
